Introduce a flag to determine whether to replace env var or not
